# Feature/aasequence speed