--- a/src/openms/source/CHEMISTRY/ResidueDB.cpp
+++ b/src/openms/source/CHEMISTRY/ResidueDB.cpp
@@ -514,11 +514,22 @@ namespace OpenMS
     bool residue_found(true), mod_found(true);
     #pragma omp critical (ResidueDB)
     {
-      if (residue_names_.find(res_name) == residue_names_.end())
+      // Perform a single lookup of the residue name in our database, we assume
+      // that if it is present in residue_mod_names_ then we have seen it
+      // before and can directly grab it. If its not present, we may have as
+      // unmodified residue in residue_names_ but need to create a new entry as
+      // modified residue. If the residue itself is unknow, we will throw (see
+      // below).
+      const auto& rm_entry = residue_mod_names_.find(res_name);
+      if (rm_entry == residue_mod_names_.end())
       {
-        residue_found = false;
+        if (residue_names_.find(res_name) == residue_names_.end())
+        {
+          residue_found = false;
+        }
       }
-      else
+
+      if (residue_found)
       {
         const ResidueModification* mod;
         try
@@ -537,11 +548,17 @@ namespace OpenMS
           const String& id = mod->getId().empty() ? mod->getFullId() : mod->getId();
 
           // check if modified residue is already present in ResidueDB
-          if (residue_mod_names_.has(res_name) && residue_mod_names_[res_name].has(id))
+          bool found = false;
+          if (rm_entry != residue_mod_names_.end())
           {
-            res = residue_mod_names_[res_name][id];
+            const auto& inner = rm_entry->second.find(id);
+            if (inner != rm_entry->second.end())
+            {
+              res = inner->second;
+              found = true;
+            }
           }
-          else
+          if (!found)
           {
             // create and register this modified residue
             res = new Residue(*residue_names_[res_name]);