Several performance improvements for `AASequence` and parsing from string, especially for modified sequences

- use `std::unordered_map` to store modification names
- reserve space for new residues
- faster check if residues match a given modification or not
- remove unnecessary double lookups in the map

# Benchmarks

Benchmarking was done with 

```
  int nr_mol = 2e7; // for AASequence
  nr_mol = 1e6; // for Oxidation
  int nr_fragments = 0;
  for (Size i = 0; i < nr_mol;  i++) 
  {
    auto pep = AASequence::fromString("DFPMIAMNGMER"); 
    nr_fragments += pep.size();
  }
```

## Unmodified AASequence

This is quite fast, the current PR reduces execution time from 4.5 s to 2s for 2e7 iterations, basically reducing time from 220 nsec to 95 nsec for parsing. This is mainly due to reserving space in the vector

## Modified AASequence

Modifications slow down parsing significantly, taking about 3.2 usec before optimization. Therefore a modified sequence would take about 15x longer to parse (or 32x longer after the optimization above) which means that speed will always be dominated by modified AA strings. A modification with 3 oxidized residues now takes only 1.5 usec to parse (down from 3.2 usec) with each modification taking about 400 nsec to parse:

- 1.5 usec for three mods
- 1.1 usec for two mods
- 700 nsec for one modification

Most of the time is spent looking up modifications names in `searchModifications` and in identifying the correct one among the results (e.g. for Oxidation there are 13 results in the hash, so that means the code needs to iterate through all 13 and find the correct one. It seems a bit excessive since we have a different modification for each oxidized amino acid).