@hroest 2020-01-07
Another issue is that we have almost 10k different modification names that we store which also makes this highly inefficient. This is probably why the lookup is so slow.