@hroest 2020-01-10
speed is now down from initially 3.2 usec to previously 1.5 usec to now 750 nsec for the triple modification peptide. The time it takes to parse the modified peptide `DFPM(Oxidation)IAM(Oxidation)NGM(Oxidation)ER` is as follows:

90 nsec if it has any bracket + 190 * x nanoseconds (x = nr mods)

or

```
105 + has_mod * 90 + nr_mods * 190
```

So overall from the starting point of 3.2 usec, this is a speedup of 4-fold for parsing modified peptides. It also means that peptides with a single modification will take around 400 nsec to parse which is in the same order of magnitude than unmodified peptides (100 nsec) and actually similarly long as an unmodified peptide took before the optimization.