@timosachsenberg 2020-01-14
I did not check all details but I guess everything is still thread-safe?