@hroest 2020-01-16
> I did not check all details but I guess everything is still thread-safe?

yes, we are actually testing that using a multithreaded test and if we remove the openmp `pragma` then it fails. The new function also has the pragmas