@timosachsenberg 2020-02-20
looks good from my side. maybe add some comments and consider to fix the style suggestions